Adicionando o recurso de edição
